# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certificate Manager API",
        "description": "API REST unificada (v1) para la gestión completa de solicitudes de certificados digitales: creación, archivos adjuntos, emisión agnóstica de proveedor (correo legacy / Viafirma RA PKCS#10 Zero-Touch), pagos WOMPI, cuotas POSTPAID, analíticas IA, notificaciones de vencimiento y webhooks. Requiere autenticación OAuth 2.0 con Laravel Passport.",
        "contact": {
            "name": "Soporte Matias",
            "email": "soporte@matias.com.co"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 — única versión soportada"
        }
    ],
    "paths": {
...
        "/certificate-request/{id}/issue": {
            "post": {
                "tags": [
                    "Emisión de Certificados"
                ],
                "summary": "Disparar emisión del certificado (provider-agnostic)",
                "description": "Inicia el flujo de emisión usando el proveedor activo (mail / viafirma / futuros). El proveedor se resuelve por cascada: payload override (admin) → companies.issuance_provider → CERTIFICATE_ISSUANCE_PROVIDER → fallback 'mail'. Para Viafirma el cuerpo debe incluir 'email_certificate'.",
                "operationId": "certificateRequestIssue",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud (certificate_requests.id)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IssueCertificateBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Emisión por correo (mail) ejecutada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssuanceResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Solicitud Viafirma creada (estado SUBMITTED)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssuanceResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ya existe un trámite Viafirma activo para esta solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida o proveedor no soporta la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Error transitorio del proveedor remoto (Viafirma)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/certificate-request/{id}/send-mail": {
            "post": {
                "tags": [
                    "Emisión de Certificados"
                ],
                "summary": "[DEPRECATED] Enviar solicitud por correo electrónico",
                "description": "Alias de compatibilidad con el flujo legacy. Internamente delega al proveedor 'mail'. Será eliminado tras el sunset window (header Sunset). Migrar a POST /certificate-request/{id}/issue.",
                "operationId": "certificateRequestSendMailDeprecated",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comments": {
                                        "type": "string",
                                        "example": "Solicitud reprocesada",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Correo enviado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssuanceResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No fue posible enviar la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "deprecated": true,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/certificate-request/{id}/issuance": {
            "get": {
                "tags": [
                    "Emisión de Certificados"
                ],
                "summary": "Consultar estado del trámite",
                "description": "Devuelve el estado normalizado del trámite de emisión. Para Viafirma incluye internal_state, remote_status, public_id y fechas relevantes. Para 'mail' devuelve el request_status actual.",
                "operationId": "certificateRequestIssuanceShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del trámite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssuanceResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud (o trámite Viafirma) no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/certificate-request/{id}/issuance/download": {
            "get": {
                "tags": [
                    "Emisión de Certificados"
                ],
                "summary": "Metadata de descarga del P12 (sólo Viafirma)",
                "description": "Retorna el PIN temporal y una URL firmada (24h) para descargar el .p12 ensamblado. Disponible cuando el estado interno es ASSEMBLED o COMPLETED.",
                "operationId": "certificateRequestIssuanceDownload",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadata + PIN",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssuanceDownloadMetadata"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trámite no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "El estado actual no permite descarga / proveedor no soporta descarga",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "El PIN del certificado fue purgado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/certificate-request/{id}/issuance/download/file": {
            "get": {
                "tags": [
                    "Emisión de Certificados"
                ],
                "summary": "Descarga binaria del P12 (streaming)",
                "description": "Streaming directo del archivo .p12 con Content-Disposition: attachment. Sólo proveedor Viafirma en estado ASSEMBLED/COMPLETED.",
                "operationId": "certificateRequestIssuanceDownloadFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Binario P12",
                        "content": {
                            "application/x-pkcs12": {}
                        }
                    },
                    "404": {
                        "description": "Trámite no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "El estado actual no permite descarga / proveedor no soporta descarga",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Descripción del error"
                    }
                },
                "type": "object"
            },
...
            "IssueCertificateBody": {
                "description": "Payload del endpoint unificado de emisión: POST /certificate-request/{id}/issue. Todos los campos son opcionales — el sistema resuelve el proveedor activo (mail/viafirma) por config + reglas de la empresa.",
                "properties": {
                    "provider": {
                        "description": "Override del proveedor. Sólo se honra si el caller es admin y la config CERTIFICATE_ISSUANCE_ALLOW_OVERRIDE está activa.",
                        "type": "string",
                        "enum": [
                            "mail",
                            "viafirma"
                        ],
                        "example": "viafirma",
                        "nullable": true
                    },
                    "email_certificate": {
                        "description": "Email de notificación KYC. OBLIGATORIO cuando el provider resuelto/forzado es 'viafirma'.",
                        "type": "string",
                        "format": "email",
                        "example": "rep.legal@empresa.com",
                        "nullable": true
                    },
                    "organization_type": {
                        "description": "Tipo de organización Viafirma. Obligatorio para Persona Jurídica.",
                        "type": "string",
                        "enum": [
                            "RM",
                            "PROP",
                            "RUNEOL",
                            "RNT",
                            "ESAL",
                            "ESOL",
                            "JUEGOS",
                            "EXTRANJERAS"
                        ],
                        "example": "RM",
                        "nullable": true
                    },
                    "identity_type_override": {
                        "description": "Override del tipo de identidad del solicitante. Si null se deriva del catálogo.",
                        "type": "string",
                        "enum": [
                            "IDC",
                            "PAS"
                        ],
                        "example": null,
                        "nullable": true
                    },
                    "comments": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Solicitud expedita por convenio empresarial",
                        "nullable": true
                    },
                    "metadata": {
                        "description": "Metadatos libres adjuntos a la auditoría",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "IssuanceResultData": {
                "description": "Resultado normalizado de cualquier proveedor de emisión (mail, viafirma, ...).",
                "properties": {
                    "provider": {
                        "type": "string",
                        "enum": [
                            "mail",
                            "viafirma"
                        ],
                        "example": "viafirma"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "sent",
                            "submitted",
                            "processing",
                            "ready",
                            "completed",
                            "failed",
                            "unsupported"
                        ],
                        "example": "submitted"
                    },
                    "message": {
                        "type": "string",
                        "example": "Solicitud Viafirma creada exitosamente."
                    },
                    "external_id": {
                        "description": "ID externo del proveedor (cod_request en Viafirma).",
                        "type": "string",
                        "example": "PYJR5N4QC",
                        "nullable": true
                    },
                    "resource_id": {
                        "description": "ID interno del agregado de emisión.",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "data": {
                        "description": "Payload extendido específico del proveedor. Ejemplo Viafirma: public_id, internal_state, remote_status.",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "IssuanceResponse": {
                "description": "Envoltura estándar de los endpoints de emisión.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Solicitud Viafirma creada exitosamente."
                    },
                    "data": {
                        "$ref": "#/components/schemas/IssuanceResultData"
                    }
                },
                "type": "object"
            },
            "IssuanceDownloadMetadata": {
                "description": "Metadata de descarga del P12 (sólo proveedor Viafirma cuando el estado interno es ASSEMBLED/COMPLETED).",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "p12_pin": {
                        "description": "PIN temporal para abrir el archivo P12.",
                        "type": "string",
                        "example": "X3kP9aQ1mZv7"
                    },
                    "p12_filename": {
                        "type": "string",
                        "example": "empresa_900400300.p12"
                    },
                    "download_url": {
                        "description": "URL firmada temporal (24h) para descarga directa. Null si el disco no soporta URLs firmadas.",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-20T18:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de acceso OAuth 2.0 obtenido desde /authentication/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}